Eliminate McpMessageService and consolidate orchestration into McpCommunicationService
- Removed redundant McpMessageService abstraction layer
- Moved MCP request routing logic directly into McpCommunicationService
- Updated Program.cs DI registrations to remove McpMessageService
- Simplified architecture by reducing unnecessary service dependencies
